fix(entities): change fileSize from Int to Long to support files > 2GB
Telegram's file_size field can exceed Int.MAX_VALUE (2,147,483,647).
Parsing large file sizes as Int causes a NumberFormatException crash.

+++ telegram/src/main/kotlin/com/github/kotlintelegrambot/entities/files/Audio.kt
@@ -13,7 +13,7 @@ data class Audio(
     @SerializedName(FilesFields.performer) val performer: String? = null,
     @SerializedName(FilesFields.title) val title: String? = null,
     @SerializedName(FilesFields.mimeType) val mimeType: String? = null,
-    @SerializedName(FilesFields.fileSize) val fileSize: Int? = null,
+    @SerializedName(FilesFields.fileSize) val fileSize: Long? = null,
     @SerializedName(FilesFields.thumb) val thumb: PhotoSize? = null,
     @SerializedName(FilesFields.fileName) val fileName: String? = null,
 )
